refactor: move to inlang sdk v2
