+++ inlang/source-code/plugins/i18next/package.json
@@ -21,24 +21,21 @@
 		"clean": "rm -rf ./dist ./node_modules"
 	},
 	"dependencies": {
-		"@inlang/detect-json-formatting": "workspace:*",
-		"@inlang/sdk": "workspace:*",
+		"@inlang/sdk2": "workspace:*",
 		"@sinclair/typebox": "0.31.28"
 	},
 	"devDependencies": {
 		"@inlang/cli": "workspace:*",
-		"@inlang/sdk": "workspace:*",
-		"@inlang/plugin": "workspace:*",
-		"@lix-js/fs": "workspace:*",
 		"@size-limit/preset-small-lib": "^8.2.4",
 		"@types/flat": "^5.0.2",
 		"@types/lodash.merge": "^4.6.7",
 		"@types/parsimmon": "1.10.6",
-		"@vitest/coverage-v8": "^0.33.0",
+		"@vitest/coverage-v8": "^2.0.5",
 		"flat": "^5.0.2",
 		"parsimmon": "^1.18.1",
 		"patch-package": "6.5.1",
+		"prettier": "^3.3.3",
 		"typescript": "^5.5.2",
-		"vitest": "0.33.0"
+		"vitest": "^2.0.5"
 	}
 }
